requirements.txt
fastapi==0.65.1
starlette==0.14.2
pydantic==1.8.2
uvicorn==0.13.4
tweepy==3.10.0
fire==0.4.0
wandb==0.10.28
python-dotenv==0.17.0
transformers==4.5.1
pymongo==3.10.1
python-multipart==0.0.5